feat: add max length for expression; refactor utils folder
