+++ src/utils/helpers/checkCorrectBrakcets.js
