+++ src/utils/helpers/checkCorrectOperators.js
