+++ src/utils/helpers/checkPriority.js
