+++ src/utils/helpers/cutZeros.js
