+++ src/utils/helpers/deleteLastItem.js
