+++ src/utils/helpers/doCorrectValue.js
