+++ src/utils/helpers/removeItemFromNumberStack.js
