+++ src/utils/helpers/removeItemFromOperatorStack.js
