+++ src/utils/helpers/replacePreviousOperator.js
